// src/types/inventories.ts
export interface ItemStatBoost {
  name: string;
  percentage: number;
  slug: string;
}

export interface ItemType {
  name: string;
  slug: string;
  sort_order?: number;
  stackable?: number | boolean;
}

export interface Medical {
  agility_boost: number;
  base_toxicity: number;
  base_toxicity_formatted: string;
  intelligence_boost: number;
  social_boost: number;
  stamina_boost: number;
  strength_boost: number;
}

export interface Armor {
  energy: number;
  impact: number;
  piercing: number;
}

export interface WeaponType {
  is_long_range: boolean;
  name: string;
}

export interface Weapon {
  accuracy: number;
  energy_damage: number;
  hand_to_hand: boolean;
  impact_damage: number;
  piercing_damage: number;
  weapon_type: WeaponType;
}

export interface InventoryItemDetails {
  bonds?: number;
  description: string;
  image: string;
  item_component_armor?: Armor;
  item_component_medical?: Medical;
  item_component_mod?: object;
  item_component_stats?: ItemStatBoost[];
  item_component_weapon?: Weapon;
  item_type: ItemType;
  mass: number;
  name: string;
  rarity: string;
  stack_size: number;
  slot_type: string;
  slug: string;
  tier: number;
}

export interface PriceRange {
  minimum_price: number;
  maximum_price: number;
}

interface ActionMessage {
  message: string;
}

export type Modifiers = "reputation" | "social";

export interface Modifier {
  amount: number;
  is_good: boolean;
  modifier: number;
  percentage: string;
  reason: string;
  sign: string;
  slug: string;
  type: string;
}

export interface PriceModifiers {
  base_price: Price;
  final_price: Price;
  price_modifiers: Modifier[];
}

export interface ItemActionOptions {
  bonds_fee?: Price;
  credits_fee_multiplier?: number;
  min_credits_fee?: number;
  pricing?: PriceModifiers;
  duration?: string;
  max_purchase_quantity?: string;
}

export interface ItemAction {
  enabled: boolean;
  max_purchase_quantity: string;
  messages: ActionMessage[];
  name: string;
  options: ItemActionOptions;
  slug: string;
  url: string;
}
// ...
export type item_type =
  | "armor"
  | "weapon"
  | "blueprint"
  | "medical"
  | "trade-good"
  | "mission"
  | "ration"
  | "food"
  | "event";

export type rarity = "common" | "uncommon" | "rare" | "epic";